+++ R/aftables-package.R
@@ -6,13 +6,12 @@
 ## usethis namespace: start
 #' @import openxlsx2
 #' @importFrom dplyr across mutate everything filter all_of
-#' @importFrom tidyr replace_na
-#' @importFrom scales number
 #' @importFrom utils file.edit
 #' @importFrom yaml read_yaml
 #' @importFrom rlang .data
 #' @importFrom pillar tbl_sum
 #' @importFrom purrr pwalk
+#' @importFrom purrr pluck
 #' @importFrom stringr str_extract
 #' @importFrom stringr str_replace
 #' @importFrom stringr str_replace_all